Move import logic out of index class [skip ci]
